[Ref] Convert 32-bit ModInstrument members to 16-bit. This is more than enough for the data held by these members, and helps with future bug fixes that require some of these members to be copied to the already way-too-large ModChannel struct.
git-svn-id: https://source.openmpt.org/svn/openmpt/trunk/OpenMPT@24774 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ mptrack/View_pat.cpp
@@ -3067,7 +3067,7 @@ std::optional<int> CViewPattern::GetDefaultVolume(const ModCommand &m, ModComman
 			return std::min(sample.nVolume, uint16(256)) / 4u;
 	} else if(m.instr > 0 && m.instr <= sndFile.GetNumInstruments() && sndFile.Instruments[m.instr] != nullptr && sndFile.Instruments[m.instr]->HasValidMIDIChannel())
 	{
-		return std::min(sndFile.Instruments[m.instr]->nGlobalVol, uint32(64));  // For instrument plugins
+		return std::min(sndFile.Instruments[m.instr]->nGlobalVol, uint16(64));  // For instrument plugins
 	}
 	return std::nullopt;
 }
